Move files out of server/resources/Product directory
