+++ server/index.js
@@ -5,7 +5,7 @@ const path = require('path');
 const compression = require('compression');
 
 // Controller Functions
-const reviewController = require('./resources/Product/reviewController.js');
+const reviewController = require('./reviewController.js');
 
 // Configuration
 const port = process.env.PORT || 3013;
@@ -21,8 +21,6 @@ app.use('/', express.static(path.join(__dirname, '../client/dist')));
 
 // API Endpoints
 app.use('/api/reviews/:number', reviewController.retrieveOne);
-
-module.exports = app;
 
 app.listen(port, () => {
   console.log(`Reviews component server listing on port# ${port}`);
